+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [2.2.0] - 2020-08-xx
+## [2.2.0] - 2020-08-06
 
 - [#28](https://github.com/estruyf/vscode-front-matter/issues/28): Align the file its name with the article slug
 - [#47](https://github.com/estruyf/vscode-front-matter/issues/47): Fix when table shows only value `0`